Initial commit of mostly-complete first KirHut Weekend Project.
This version lacks much in the way of translations and Internationalization support but that will be fixed eventually. This is just meant to be an example of how to get a simple Dice Rolling app working.

Implemented rolling, history, clearing dice and history, and a few other bits and bobs. Icons are from this free icon source: https://opengameart.org/content/hit-die-icons

This also gives an example of very simple use of the Qt Resources system which I quite like.

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>