Several bugfixes and added full translations to Spanish, French, and Korean
Mostly just got the translation system working, and modified the main.cpp file to allow passing in localities by command line so it was easy to test them. The application can probably do with some more work on command line input or moving the History widget and making the interface sharper, but I kind of doubt that will happen. This looks pretty good for what I was going for even if the History screen is janky and not that great.

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>